+++ docs/effect設置格式.lua
@@ -1,6 +1,7 @@
 setting = {
     (要跟檔名一樣)name,
     target,
+    (可選)source,
     (可選)value{number, percent},
     (可選)time,
 }